pyproject.toml
[project]
dependencies = [
    "loguru>=0.6.0",
    "python-telegram-bot>=13.13",
    "envparse>=0.2.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
